backend/requirements.txt
channels==4.0.0
Django==5.0.3
djangorestframework==3.14.0
openai==1.13.3
python-dotenv==1.0.1
django_ethereum==0.0.4
web3==6.17.0
exponent_server_sdk==2.1.0
daphne==4.1.2
channels_redis==4.2.0
coverage==7.4.4
pytest-django==4.8.0
pytest-cov==5.0.0
psycopg2-binary==2.9.9
pytest-asyncio==0.23.6
django-cors-headers==4.3.1
